+++ test/fixtures/remove_extension.fixture.ts
