+++ test/fixtures/strip_string.fixture.ts
